Rename `embeddings.py` to `models.py` in-namespace
